Conflicts:
	Gemfile
	Gemfile.lock
	Rakefile
	_config.yml

+++ Rakefile
@@ -1,5 +1,6 @@
 require "rubygems"
 require "bundler/setup"
+require "stringex"
 
 ## -- Rsync Deploy config -- ##
 # Be sure your public key is listed in your server's ~/.ssh/authorized_keys file
@@ -11,20 +12,25 @@ deploy_default = "push"
 deploy_branch  = "master"
 deploy_url     = "git@github.com:Gonzih/gonzih.github.com.git"
 
-## -- Misc Configs, you probably have no reason to changes these -- ##
+## -- Misc Configs -- ##
 
-public_dir   = "public"    # compiled site directory
-source_dir   = "source"    # source file directory
-deploy_dir   = "_deploy"   # deploy directory (for Github pages deployment)
-stash_dir    = "_stash"    # directory to stash posts for speedy generation
-posts_dir    = "_posts"    # directory for blog files
-themes_dir   = ".themes"   # directory for blog files
-new_post_ext = "markdown"  # default new post file extension when using the new_post task
-new_page_ext = "markdown"  # default new page file extension when using the new_page task
+public_dir      = "public"    # compiled site directory
+source_dir      = "source"    # source file directory
+blog_index_dir  = 'source'    # directory for your blog's index page (if you put your index in source/blog/index.html, set this to 'source/blog')
+deploy_dir      = "_deploy"   # deploy directory (for Github pages deployment)
+stash_dir       = "_stash"    # directory to stash posts for speedy generation
+posts_dir       = "_posts"    # directory for blog files
+themes_dir      = ".themes"   # directory for blog files
+new_post_ext    = "markdown"  # default new post file extension when using the new_post task
+new_page_ext    = "markdown"  # default new page file extension when using the new_page task
+server_port     = "4000"      # port for preview server eg. localhost:4000
 
 
 desc "Initial setup for Octopress: copies the default theme into the path of Jekyll's generator. Rake install defaults to rake install[classic] to install a different theme run rake install[some_theme_name]"
 task :install, :theme do |t, args|
+  if File.directory?(source_dir) || File.directory?("sass")
+    abort("rake aborted!") if ask("A theme is already installed, proceeding will overwrite existing files. Are you sure?", ['y', 'n']) == 'n'
+  end
   # copy theme into working Jekyll directories
   theme = args.theme || 'classic'
   puts "## Copying "+theme+" theme into ./#{source_dir} and ./sass"
@@ -42,30 +48,57 @@ end
 
 desc "Generate jekyll site"
 task :generate do
+  raise "### You haven't set anything up yet. First run `rake install` to set up an Octopress theme." unless File.directory?(source_dir)
   puts "## Generating Site with Jekyll"
   system "jekyll"
 end
 
 desc "Watch the site and regenerate when it changes"
 task :watch do
-  system "trap 'kill $jekyllPid $compassPid' Exit; jekyll --auto & jekyllPid=$!; compass watch & compassPid=$!; wait"
+  raise "### You haven't set anything up yet. First run `rake install` to set up an Octopress theme." unless File.directory?(source_dir)
+  puts "Starting to watch source with Jekyll and Compass."
+  jekyllPid = Process.spawn("jekyll --auto")
+  compassPid = Process.spawn("compass watch")
+
+  trap("INT") {
+    [jekyllPid, compassPid].each { |pid| Process.kill(9, pid) rescue Errno::ESRCH }
+    exit 0
+  }
+
+  [jekyllPid, compassPid].each { |pid| Process.wait(pid) }
 end
 
 desc "preview the site in a web browser"
 task :preview do
-  system "trap 'kill $jekyllPid $compassPid' Exit; jekyll --auto --server & jekyllPid=$!; compass watch & compassPid=$!; wait"
+  raise "### You haven't set anything up yet. First run `rake install` to set up an Octopress theme." unless File.directory?(source_dir)
+  puts "Starting to watch source with Jekyll and Compass. Starting Rack on port #{server_port}"
+  jekyllPid = Process.spawn("jekyll --auto")
+  compassPid = Process.spawn("compass watch")
+  rackupPid = Process.spawn("rackup --port #{server_port}")
+
+  trap("INT") {
+    [jekyllPid, compassPid, rackupPid].each { |pid| Process.kill(9, pid) rescue Errno::ESRCH }
+    exit 0
+  }
+
+  [jekyllPid, compassPid, rackupPid].each { |pid| Process.wait(pid) }
 end
 
 # usage rake new_post[my-new-post] or rake new_post['my new post'] or rake new_post (defaults to "new-post")
 desc "Begin a new post in #{source_dir}/#{posts_dir}"
 task :new_post, :title do |t, args|
+  raise "### You haven't set anything up yet. First run `rake install` to set up an Octopress theme." unless File.directory?(source_dir)
   require './plugins/titlecase.rb'
+  mkdir_p "#{source_dir}/#{posts_dir}"
   args.with_defaults(:title => 'new-post')
   title = args.title
-  filename = "#{source_dir}/#{posts_dir}/#{Time.now.strftime('%Y-%m-%d')}-#{title.downcase.gsub(/&/,'and').gsub(/[,'":\?!\(\)\[\]]/,'').gsub(/[\W\.]/, '-').gsub(/-+$/,'')}.#{new_post_ext}"
+  filename = "#{source_dir}/#{posts_dir}/#{Time.now.strftime('%Y-%m-%d')}-#{title.to_url}.#{new_post_ext}"
+  if File.exist?(filename)
+    abort("rake aborted!") if ask("#{filename} already exists. Do you want to overwrite?", ['y', 'n']) == 'n'
+  end
   puts "Creating new post: #{filename}"
   open(filename, 'w') do |post|
-    system "mkdir -p #{source_dir}/#{posts_dir}";
+    system "mkdir -p #{source_dir}/#{posts_dir}/";
     post.puts "---"
     post.puts "layout: post"
     post.puts "title: \"#{title.gsub(/&/,'&amp;').titlecase}\""
@@ -79,6 +112,7 @@ end
 # usage rake new_page[my-new-page] or rake new_page[my-new-page.html] or rake new_page (defaults to "new-page.markdown")
 desc "Create a new page in #{source_dir}/(filename)/index.#{new_page_ext}"
 task :new_page, :filename do |t, args|
+  raise "### You haven't set anything up yet. First run `rake install` to set up an Octopress theme." unless File.directory?(source_dir)
   require './plugins/titlecase.rb'
   args.with_defaults(:filename => 'new-page')
   page_dir = source_dir
@@ -89,6 +123,9 @@ task :new_page, :filename do |t, args|
     filename = "#{name}.#{extension}"
     mkdir_p page_dir
     file = page_dir + filename
+    if File.exist?(file)
+      abort("rake aborted!") if ask("#{file} already exists. Do you want to overwrite?", ['y', 'n']) == 'n'
+    end
     puts "Creating new page: #{file}"
     open(file, 'w') do |page|
       page.puts "---"
@@ -122,7 +159,7 @@ end
 
 desc "Clean out caches: _code_cache, _gist_cache, .sass-cache"
 task :clean do
-  system "rm -rf _code_cache/** _gist_cache/** .sass-cache/**"
+  rm_rf ["_code_cache/**", "_gist_cache/**", ".sass-cache/**", "source/stylesheets/screen.css"]
 end
 
 desc "Move sass to sass.old, install sass theme updates, replace sass/custom with sass.old/custom"
@@ -130,11 +167,11 @@ task :update_style, :theme do |t, args|
   theme = args.theme || 'classic'
   if File.directory?("sass.old")
     puts "removed existing sass.old directory"
-    system "rm -r sass.old"
+    rm_r "sass.old", :secure=>true
   end
-  system "mv sass sass.old"
+  mv "sass", "sass.old"
   puts "## Moved styles into sass.old/"
-  system "mkdir -p sass; cp -R #{themes_dir}/"+theme+"/sass/ sass/"
+  cp_r "#{themes_dir}/"+theme+"/sass/", "sass"
   cp_r "sass.old/custom/.", "sass/custom"
   puts "## Updated Sass ##"
 end
@@ -143,14 +180,15 @@ desc "Move source to source.old, install source theme updates, replace source/_i
 task :update_source, :theme do |t, args|
   theme = args.theme || 'classic'
   if File.directory?("#{source_dir}.old")
-    puts "removed existing #{source_dir}.old directory"
-    system "rm -r #{source_dir}.old"
+    puts "## Removed existing #{source_dir}.old directory"
+    rm_r "#{source_dir}.old", :secure=>true
   end
-  system "mv #{source_dir} #{source_dir}.old"
-  puts "moved #{source_dir} into #{source_dir}.old/"
-  system "mkdir -p #{source_dir}; cp -R #{themes_dir}/"+theme+"/source/. #{source_dir}"
-  system "cp -Rn #{source_dir}.old/. #{source_dir}"
-  system "cp -Rf #{source_dir}.old/_includes/custom/. #{source_dir}/_includes/custom/"
+  cp_r "#{source_dir}/.", "#{source_dir}.old"
+  puts "## Copied #{source_dir} into #{source_dir}.old/"
+  cp_r "#{themes_dir}/"+theme+"/source/.", source_dir, :remove_destination=>true
+  cp_r "#{source_dir}.old/_includes/custom/.", "#{source_dir}/_includes/custom/", :remove_destination=>true
+  mv "#{source_dir}/index.html", "#{blog_index_dir}", :force=>true if blog_index_dir != source_dir
+  cp "#{source_dir}.old/index.html", source_dir if blog_index_dir != source_dir
   puts "## Updated #{source_dir} ##"
 end
 
@@ -159,7 +197,23 @@ end
 ##############
 
 desc "Default deploy task"
-task :deploy => "#{deploy_default}" do
+task :deploy do
+  [:copydot, "#{deploy_default}"].each { |t| Rake::Task[t].execute }
+end
+
+desc "Generate website and deploy"
+task :gen_deploy do
+  [:integrate, :generate, :deploy].each { |t| Rake::Task[t].execute }
+end
+
+desc "copy dot files for deployment"
+task :copydot do
+  exclusions = [".", "..", ".DS_Store"]
+  Dir["#{source_dir}/**/.*"].each do |file|
+    if !File.directory?(file) && !exclusions.include?(file)
+      cp(file, file.gsub(/#{source_dir}/, "#{public_dir}"));
+    end
+  end
 end
 
 desc "Deploy website via rsync"
@@ -169,8 +223,7 @@ task :rsync do
 end
 
 desc "deploy public directory to github pages"
-task :push do
-  system "git clone #{deploy_url} #{deploy_dir}" unless Dir.exist? deploy_dir
+multitask :push do
   puts "## Deploying branch to Github Pages "
   (Dir["#{deploy_dir}/*"]).each { |f| rm_rf(f) }
   system "cp -R #{public_dir}/* #{deploy_dir}"
@@ -251,8 +304,22 @@ def ok_failed(condition)
   end
 end
 
+def get_stdin(message)
+  print message
+  STDIN.gets.chomp
+end
+
+def ask(message, valid_options)
+  if valid_options
+    answer = get_stdin("#{message} #{valid_options.to_s.gsub(/"/, '').gsub(/, /,'/')} ") while !valid_options.include?(answer)
+  else
+    answer = get_stdin(message)
+  end
+  answer
+end
+
 desc "list tasks"
 task :list do
-  puts "Tasks: #{(Rake::Task.tasks - [Rake::Task[:list]]).to_sentence}"
+  puts "Tasks: #{(Rake::Task.tasks - [Rake::Task[:list]]).join(', ')}"
   puts "(type rake -T for more detail)\n\n"
 end
